Break up form gallery into multiple files
