+++ jsapp/js/components/formGallery/formGallery.component.scss
@@ -3,9 +3,6 @@
 .gallery {
   background-color: colors.$kobo-white;
 }
-body {
-  margin-top: 99px !important;
-}
 .gallery__wrapper {
   margin-top: 20px;
   margin-left: 54px;
