+++ jsapp/js/components/formGallery/formGallery.interfaces.ts
@@ -1,0 +1,2 @@
+/** Represents a JavaScript object that was parsed from JSON */
+export type Json = null | boolean | number | string | Json[] | {[key: string]: Json};
